+++ game-abioticfactor/CHANGELOG.md
@@ -6,7 +6,10 @@
 
 ## [0.2.0] - 2026-02-03
 
-- Improved: Made UE4SS Scripts, UE4SS DLL, LogicMods, and Root Folder mod installers case-insensitive to folder names
+- Improved: Made UE4SS DLL, LogicMods, and Root Folder mod installers case-insensitive to folder names
+- Fixed: path strings
+- Fixed: Xbox game version detection
+- Added: Buttons to open UE4SS settings files, PCGamingWiki page, view changelog, and submit bug reports
 
 ## [0.1.1] - 2025-12-15